+++ src/pages/Soomin/Login/SoominLogin.js
@@ -1,40 +1,52 @@
-import React from 'react';
 import { useState } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import './SoominLogin.scss';
 
 export default function SoominLogin() {
-  const navigate = useNavigate();
-  const [id, setId] = useState('');
-  const [password, setPassword] = useState('');
+  const [inputValues, setInputValues] = useState({ email: '', password: '' });
 
   const goToMain = () => {
-    navigate('/soomin-main');
-  };
-  const saveUserId = event => {
-    setId(event.target.value);
-  };
-  const saveUserPassword = event => {
-    setPassword(event.target.value);
+    fetch('http://10.58.52.134:3000/users/login', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json;charset=utf-8',
+      },
+      body: JSON.stringify({
+        email: inputValues.email,
+        password: inputValues.password,
+      }),
+    })
+      .then(response => response.json())
+      .then(data => {
+        localStorage.setItem('token', data.access.Token);
+      });
   };
 
-  const inputValue = id.includes('@') && password.length >= 5;
+  const handleInput = event => {
+    const { name, value } = event.target;
+    setInputValues({ ...inputValues, [name]: value });
+  };
+
+  const inputValue =
+    inputValues.email.includes('@') && inputValues.password.length >= 4;
 
   return (
     <div className="soominlogin">
       <h1 className="logo">westagram</h1>
       <main className="user-info">
         <input
+          name="email"
           className="info-box"
           type="text"
           placeholder="전화번호, 사용자 이름 또는 이메일"
-          onChange={saveUserId}
+          onChange={handleInput}
         />
         <input
+          name="password"
           className="info-box"
           type="password"
           placeholder="비밀번호"
-          onChange={saveUserPassword}
+          onChange={handleInput}
         />
         <button
           id="login-button"
@@ -47,9 +59,17 @@ export default function SoominLogin() {
         </button>
       </main>
       <footer className="login-footer">
-        <Link className="find-password" to="/">
-          비밀번호를 잊으셨나요?
-        </Link>
+        <div className="footer-box">
+          <Link className="find-password" to="/">
+            비밀번호를 잊으셨나요?
+          </Link>
+          <span className="sign-up-box">
+            계정이 없으신가요?
+            <button className="sign-up" type="button">
+              회원가입
+            </button>
+          </span>
+        </div>
       </footer>
     </div>
   );
