+++ src/pages/Soomin/Login/SoominLogin.scss
@@ -53,12 +53,29 @@ body {
   }
 
   .login-footer {
-    margin-top: 60px;
+    margin-top: 40px;
 
-    .find-password {
-      text-decoration: none;
-      font-size: 10px;
-      color: navy;
+    .footer-box {
+      display: flex;
+      justify-content: center;
+      flex-direction: column;
+
+      .find-password {
+        margin-bottom: 10px;
+        text-decoration: none;
+        font-size: 10px;
+        color: navy;
+      }
+      .sign-up-box {
+        font-size: 10px;
+
+        .sign-up {
+          margin-left: 10px;
+          font-size: 12px;
+          font-weight: 500;
+          color: rgb(97, 198, 238);
+        }
+      }
     }
   }
 }
