--- src/pages/Soomin/Main/Comment/Comment.scss
@@ -1,4 +1,0 @@
-.contents-box {
-  margin-right: 5px;
-  font-weight: 600;
-}
