+++ src/pages/Soomin/Main/Components/Comment/Comment.scss
@@ -1,0 +1,10 @@
+.contents-box {
+  margin-right: 5px;
+  font-weight: 600;
+}
+
+.heart-icon {
+  width: 8px;
+  height: 8px;
+  margin-right: 10px;
+}
