+++ src/pages/Soomin/Main/Components/Feed/Article/Article.scss
@@ -1,0 +1,174 @@
+article {
+  background-color: #fff;
+  border: 1px solid #dadada;
+}
+.from-wrapper {
+  margin: 5px;
+
+  .profile-wrapper {
+    display: flex;
+    align-items: center;
+    margin-top: 5px;
+    font-size: 12px;
+  }
+
+  .proflie-box {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin: 10px;
+  }
+
+  .user-proflie {
+    display: flex;
+    align-items: center;
+    font-weight: 600;
+
+    .profile {
+      width: 30px;
+      height: 30px;
+      border-radius: 70%;
+      overflow: hidden;
+      margin-right: 10px;
+    }
+  }
+
+  .view-more {
+    margin-left: 280px;
+  }
+}
+
+.feed-photo1 {
+  width: 400px;
+  height: 440px;
+}
+
+.feed-icon {
+  display: flex;
+  margin: 10px;
+
+  .heart-icon {
+    width: 20px;
+    height: 20px;
+    margin-right: 10px;
+  }
+
+  .comment-icon {
+    width: 18px;
+    height: 18px;
+    margin-right: 10px;
+  }
+
+  .upload-icon {
+    width: 17px;
+    height: 17px;
+  }
+
+  .bookmark-icon {
+    width: 18px;
+    height: 18px;
+    margin-left: 287px;
+  }
+}
+
+.feed-info {
+  display: flex;
+  align-items: center;
+  margin: 10px;
+  font-size: 12px;
+
+  .follower-info {
+    display: flex;
+    align-items: center;
+    font-weight: 600;
+
+    .follower-photo {
+      width: 20px;
+      height: 20px;
+      border-radius: 70%;
+      overflow: hidden;
+      margin-right: 5px;
+    }
+  }
+}
+
+.feed-box {
+  display: flex;
+  align-items: center;
+  margin: 10px;
+  margin-bottom: 0px;
+  font-size: 12px;
+
+  .feed-contents {
+    display: flex;
+
+    .user-proflie {
+      font-weight: 600;
+    }
+    .contents {
+      width: 120px;
+      margin-left: 5px;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+    }
+  }
+
+  .comment-view-more {
+    font-size: 12px;
+    color: #b1b1b1;
+  }
+}
+
+.comment-box {
+  display: flex;
+  margin-left: 10px;
+  font-size: 12px;
+
+  .comment-wrapper {
+    width: 100%;
+    .comment {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      font-size: 12px;
+
+      .comment-heart {
+        margin-right: 10px;
+        color: #b1b1b1;
+      }
+      .contents {
+        margin-left: 5px;
+      }
+    }
+  }
+}
+
+.time {
+  margin-left: 10px;
+  font-size: 12px;
+  color: gray;
+}
+
+.comment-bar {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 5px;
+  padding: 10px;
+  border-top: 1px solid #dbdbdb;
+  font-size: 12px;
+
+  .comment-input {
+    width: 100%;
+    border: none;
+    &:focus {
+      outline: none;
+    }
+  }
+
+  .upload {
+    width: 30px;
+    font-size: 12px;
+    color: skyblue;
+  }
+}
