+++ src/pages/Soomin/Main/Components/Feed/Feed.scss
@@ -1,0 +1,4 @@
+.feed {
+  display: flex;
+  flex-direction: column;
+}
