+++ src/pages/Soomin/Main/Components/MainRight/Footer/Footer.scss
