+++ src/pages/Soomin/Main/Components/MainRight/MainRight.scss
